requirements.txt
Flask~=1.1
request~=2.22
